update with full working SSR including styles and no hydration mismatches
