+++ src/client/index.tsx
@@ -1,6 +1,6 @@
 import React from "react";
-// import App from "../shared/App";
 import { hydrateRoot } from "react-dom/client";
+import "./styles/index.scss";
 import App from "./components/App";
 const rootElement = document.getElementById("root");
 if (rootElement) {
