+++ src/client/models/Interfaces.ts
@@ -3,3 +3,22 @@ import { ReactNode } from "react";
 export interface IAuthProvider {
   children: ReactNode;
 }
+
+export interface User {
+  id: number;
+  email: string;
+}
+
+export interface AuthState {
+  user: User | null;
+  accessToken: string | null;
+  isAuthenticated: boolean;
+  loading: boolean;
+}
+
+export interface LoginCredentials {
+  email: string;
+  password: string;
+}
+
+export interface RegisterCredentials extends LoginCredentials {}
