+++ src/client/styles/index.scss
@@ -1,0 +1,1 @@
+@use '@carbon/react' with ($font-path: '@ibm/plex');
