--- src/client/types/auth.ts
@@ -1,18 +1,0 @@
-export interface User {
-  id: number;
-  email: string;
-}
-
-export interface AuthState {
-  user: User | null;
-  accessToken: string | null;
-  isAuthenticated: boolean;
-  loading: boolean;
-}
-
-export interface LoginCredentials {
-  email: string;
-  password: string;
-}
-
-export interface RegisterCredentials extends LoginCredentials {}
