+++ src/client/types/custom.d.ts
@@ -2,3 +2,5 @@ declare module "*.module.scss" {
   const classes: { [key: string]: string };
   export default classes;
 }
+
+export {};
