+++ src/client/types/global.d.ts
@@ -1,0 +1,7 @@
+declare global {
+  interface Window {
+    __INITIAL_DATA__: any;
+  }
+}
+
+export {};
