+++ src/server/render.js
@@ -1,39 +1,37 @@
 import React from "react";
 import { renderToPipeableStream } from "react-dom/server";
-import App from "../shared/App";
-// import App from "../client/components/App";
+import App from "../client/components/App";
 import { StaticRouter } from "react-router-dom";
 import { htmlEnd, htmlStart } from "../shared/htmlParts";
-import { BaseDocument } from "../shared/BaseDocument";
 
 export const render = (req, res, context, initialState) => {
-  // res.write(htmlStart) // type 1
+  res.write(htmlStart)
   const { pipe, abort } = renderToPipeableStream(
-    // type 2
-    <StaticRouter location={req.url} context={context}>
-      <BaseDocument>
-        <App initialState={initialState} />
-      </BaseDocument>
-    </StaticRouter>,
-    // type 3
-    // React.createElement(StaticRouter, { location: req.url, context }, React.createElement(App, { initialState })),
+    React.createElement(StaticRouter, { location: req.url, context }, React.createElement(App)),
     {
       bootstrapScripts: ["client.bundle.js"],
       bootstrapScriptContent: `window.__INITIAL_DATA__ = ${JSON.stringify(initialState)}`,
       onShellReady() {
-        // res.setHeader("Content-Type", "text/html");
+        if (!res.headersSent) {
+          res.setHeader('content-type', 'text/html');
+          res.status(200);
+        }
         pipe(res);
       },
       onAllReady() {
         console.log("Render: All ready");
-        // res.write(htmlEnd) // type 1
-        // res.end() // type 1
+        res.write(htmlEnd);
+        res.end();
+      },
+      onShellError(error) {
+        if (!res.headersSent) {
+          res.status(500).send("Internal server error");
+          res.statusCode = 500;
+        }
+        res.end('<h1>Something went wrong</h1>', err.message || err);
       },
       onError(err) {
         console.error('Shell error: ', err);
-        res.status(500).send("Internal server error");
-        res.statusCode = 500;
-        res.end('<h1>Something went wrong</h1>', err.message || err);
         abort();
       },
     }
