--- src/shared/App.tsx
@@ -1,10 +1,0 @@
-import React from "react";
-import { Loading } from "@carbon/react";
-
-const App = () => {
-  console.log("SERVER APP LOADING");
-
-  return <Loading active />;
-};
-
-export default App;
